Exception when trying to perform OCR
Everything else works. 
Throws an exception:
Microsoft.ProjectOxford.Vision.ClientException: Exception of type 'Microsoft.ProjectOxford.Vision.ClientException' was thrown.
I think it has to do with the picture that I am passing in, the stream is at the end of the photo. Going to try to move everything to the main class ( :/ yikes ) and call OCR on the photo there...